+++ tests/core/xmlhelp/test_core.py
@@ -28,6 +28,9 @@ def test_elementwrapper():
     with pytest.raises(KeyError, match="foo"):
         "foo" in wrapped_siddroot
 
+    with pytest.raises(KeyError, match="foo"):
+        wrapped_siddroot.get("foo")
+
     # Attribute KeyErrors
     with pytest.raises(KeyError, match="@fooattr"):
         wrapped_siddroot["@fooattr"] = "doesn't exist"
@@ -109,6 +112,22 @@ def test_elementwrapper():
     assert (
         "ECEF"
         not in wrapped_siddroot["Measurement"]["PlaneProjection"]["ReferencePoint"]
+    )
+
+    # get() defaults to empty ElementWrappers
+    assert (
+        wrapped_siddroot["Measurement"]["PlaneProjection"]["ReferencePoint"].get("ECEF")
+        == wrapped_siddroot["Measurement"]["PlaneProjection"]["ReferencePoint"]["ECEF"]
+    )
+    assert (
+        wrapped_siddroot["Measurement"]["PlaneProjection"]["ReferencePoint"].get("ECEF")
+        == {}
+    )
+    assert (
+        wrapped_siddroot["Measurement"]["PlaneProjection"]["ReferencePoint"].get(
+            "ECEF", None
+        )
+        is None
     )