+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `sarkit.cphd.Reader` support for reading a subset of vectors
 - New `check_*`methods for SICD consistency
 - More complete unit tests for SICD consistency
+- `ElementWrapper.get` now supports the `default` parameter
 
 ### Changed
 - `ElementWrapper.__contains__` now raises `KeyError` for impossible keys